Anagram Checker
Given two strings, write a function to determine if the second string is an anagram of the first. Please note that an anagram is a word, phrase or name formed by rearranging the letters of another, such as "cinema" formed from "iceman". The goal is to use the frequency counter method to solve this.